% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arrow.R
\docType{data}
\name{GeomArrow}
\alias{GeomArrow}
\title{GeomBrick}
\format{
An object of class \code{GeomArrow} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomArrow
}
\description{
GeomBrick
}
\keyword{datasets}
