% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_callout.R
\docType{data}
\name{GeomCallout}
\alias{GeomCallout}
\title{GeomBrick}
\format{
An object of class \code{GeomCallout} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomCallout
}
\description{
GeomBrick
}
\keyword{datasets}
