% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_chevron.R
\docType{data}
\name{GeomChevron}
\alias{GeomChevron}
\title{GeomBrick}
\format{
An object of class \code{GeomChevron} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomChevron
}
\description{
GeomBrick
}
\keyword{datasets}
