% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-cross.R
\docType{data}
\name{GeomCross}
\alias{GeomCross}
\title{GeomBrick}
\format{
An object of class \code{GeomCross} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomCross
}
\description{
GeomBrick
}
\keyword{datasets}
