% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-diamond.R
\docType{data}
\name{GeomDiamond}
\alias{GeomDiamond}
\title{GeomBrick}
\format{
An object of class \code{GeomDiamond} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomDiamond
}
\description{
GeomBrick
}
\keyword{datasets}
