% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-elbow.R
\docType{data}
\name{GeomElbow}
\alias{GeomElbow}
\title{GeomBrick}
\format{
An object of class \code{GeomElbow} (inherits from \code{GeomLine}, \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
GeomElbow
}
\description{
GeomBrick
}
\keyword{datasets}
