% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-parallelogram.R
\docType{data}
\name{GeomParallelogram}
\alias{GeomParallelogram}
\title{GeomBrick}
\format{
An object of class \code{GeomParallelogram} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomParallelogram
}
\description{
GeomBrick
}
\keyword{datasets}
