% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-pentagon.R
\docType{data}
\name{GeomPentagon}
\alias{GeomPentagon}
\title{GeomBrick}
\format{
An object of class \code{GeomPentagon} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomPentagon
}
\description{
GeomBrick
}
\keyword{datasets}
