% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_triangle.R
\docType{data}
\name{GeomTriangle}
\alias{GeomTriangle}
\title{GeomBrick}
\format{
An object of class \code{GeomTriangle} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomTriangle
}
\description{
GeomBrick
}
\keyword{datasets}
