% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_arrow}
\alias{make_arrow}
\title{Makes arrow}
\usage{
make_arrow(
  xmin,
  xmax,
  ymin,
  ymax,
  edge_length = 0.2,
  head_length = 0.2,
  angle = 0,
  rotate_at = "head"
)
}
\arguments{
\item{xmin}{xmin}

\item{xmax}{xmax}

\item{ymin}{ymin}

\item{ymax}{ymax}

\item{edge_length}{Edge length}

\item{head_length}{Head length}

\item{angle}{and}

\item{rotate_at}{Rotate at}
}
\description{
Makes arrow
}
