% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_cross}
\alias{make_cross}
\title{make cross}
\usage{
make_cross(xmin, xmax, ymin, ymax, wd = 0.3, angle = 0, rotate_at = "center")
}
\arguments{
\item{xmin}{xmin}

\item{xmax}{xmax}

\item{ymin}{ymin}

\item{ymax}{ymax}

\item{wd}{wd}

\item{angle}{angle}

\item{rotate_at}{rotate at}
}
\description{
make cross
}
