% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_diamond}
\alias{make_diamond}
\title{make diamond}
\usage{
make_diamond(xmin, xmax, ymin, ymax, angle = 0, rotate_at = "center")
}
\arguments{
\item{xmin}{xmin}

\item{xmax}{xmax}

\item{ymin}{ymin}

\item{ymax}{ymax}

\item{angle}{angle}

\item{rotate_at}{rotate at}
}
\description{
make diamond
}
