% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_parallelogram}
\alias{make_parallelogram}
\title{make parallelogram}
\usage{
make_parallelogram(
  xmin,
  xmax,
  ymin,
  ymax,
  angle = 0,
  theta = 30,
  rotate_at = "center"
)
}
\description{
make parallelogram
}
