% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_triangle}
\alias{make_triangle}
\title{make triangle}
\usage{
make_triangle(
  xmin,
  xmax,
  ymin,
  ymax,
  pos = 0.3,
  length = 0.2,
  angle = 0,
  rotate_at = "center"
)
}
\description{
make triangle
}
