% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_callout}
\alias{make_callout}
\title{make callout}
\usage{
make_callout(
  xmin,
  xmax,
  ymin,
  ymax,
  pos = 0.3,
  length = 0.2,
  angle = 0,
  rotate_at = "center"
)
}
\arguments{
\item{xmin}{xmin}

\item{xmax}{xmax}

\item{ymin}{ymin}

\item{ymax}{ymax}

\item{pos}{pos}

\item{length}{Length}

\item{angle}{angle}

\item{rotate_at}{rotate at}
}
\description{
make callout
}
