% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_parallelogram}
\alias{make_parallelogram}
\title{make parallelogram}
\usage{
make_parallelogram(
  xmin,
  xmax,
  ymin,
  ymax,
  angle = 0,
  theta = 15,
  rotate_at = "center"
)
}
\arguments{
\item{xmin}{xmin}

\item{xmax}{xmax}

\item{ymin}{ymin}

\item{ymax}{ymax}

\item{angle}{angle}

\item{theta}{theta}

\item{rotate_at}{rotate at}
}
\description{
make parallelogram
}
